Fix: 피드 전체타입 fetch시만 캐시 삭제하도록 수정

+++ app/src/main/java/com/android04/godfisherman/data/repository/FeedRepository.kt
@@ -28,8 +28,10 @@ class FeedRepository @Inject constructor(
     @ApplicationScope private val externalScope: CoroutineScope
 ) {
     private fun fetchPagingData(type: Type): Flow<PagingData<FeedData>> {
-        externalScope.launch {
-            localDataSource.deleteAll()
+        if (type == Type.ALL) {
+            externalScope.launch {
+                localDataSource.deleteAll()
+            }
         }
 
         return Pager(PagingConfig(pageSize = 2)) { FeedRemotePagingSource(type) }.flow
